requirements.txt
psycopg2==2.9.11
PyQt5==5.15.11
pyqt5_sip==12.17.2
python-dotenv==1.2.1
